modules/hostedfdxapi/src/main/java/com/fdx/api/FdxApiApplication.java
package com.fdx.api;

import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.boot.builder.SpringApplicationBuilder;
import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
import org.springframework.context.annotation.ComponentScan;

@SpringBootApplication
@ComponentScan(basePackages = {"com.fdx"})
public class FdxApiApplication extends SpringBootServletInitializer {

	public static void main(String[] args) {
		new SpringApplication(FdxApiApplication.class).run(args);
	}

	@Override
	protected SpringApplicationBuilder configure(SpringApplicationBuilder builder) {
		return builder.sources(FdxApiApplication.class);
	}
}